12/12/2019 - 14:11 - aluno e professor ok

+++ MeuTCC/database/migrations/2019_12_11_201430_create_tccs_table.php
@@ -14,9 +14,9 @@ class CreateTccsTable extends Migration
     public function up()
     {
         Schema::create('tccs', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('aluno_id')->unsigned();
-            $table->integer('professor_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('aluno_id')->unsigned();
+            $table->bigInteger('professor_id')->unsigned();
             $table->string('titulo', 100);
             $table->string('area');
             $table->smallInteger('ano');
